Add untested Gumbel-softmax function and note new planned architecture
I'm probably gonna try both the soft and the hard versions of this
function when generating palette indexes.
I never thought I would say this, but thanks ChatGPT for cluing me into
the existence of this function (I looked it up afterwards to make sure
it was as applicable as it sounded).